WINTERMUTE: Explicitly cast zoom for float version of TransformStruct

+++ engines/wintermute/graphics/transform_struct.cpp
@@ -49,8 +49,8 @@ TransformStruct::TransformStruct(int32 zoomX, int32 zoomY, uint32 angle, int32 h
 }
 
 TransformStruct::TransformStruct(float zoomX, float zoomY, uint32 angle, int32 hotspotX, int32 hotspotY, TSpriteBlendMode blendMode, uint32 rgbaMod, bool mirrorX, bool mirrorY, int32 offsetX, int32 offsetY) {
-	init(Point32(zoomX / 100.0 * kDefaultZoomX, 
-		     zoomY / 100.0 * kDefaultZoomY), 
+	init(Point32((int32)(zoomX / 100.0 * kDefaultZoomX), 
+		     (int32)(zoomY / 100.0 * kDefaultZoomY)), 
 		angle, 
 		Point32(hotspotX, hotspotY), 
 		false,
